+++ packages/nestjs/src/index.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const { PompelmiModule, POMPELMI_OPTIONS } = require('./pompelmi.module');
+const { PompelmiService }                  = require('./pompelmi.service');
+const { PompelmiGuard }                    = require('./pompelmi.guard');
+const { PompelmiInterceptor }              = require('./pompelmi.interceptor');
+
+module.exports = { PompelmiModule, PompelmiService, PompelmiGuard, PompelmiInterceptor, POMPELMI_OPTIONS };